requirements.txt
pandas==2.1.1
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2
huggingface-hub==0.19.0
transformers==4.35.0
tokenizers==0.15.0
fastapi==0.104.1
uvicorn==0.23.2
pydantic==1.10.13
python-multipart==0.0.6
numpy==1.26.0
gensim==4.3.2
tqdm==4.66.1
scipy==1.12.0
streamlit==1.32.0
python-dotenv==1.0.0
